+++ example/lib/main.dart
@@ -7,7 +7,7 @@ import 'routes/app_pages.dart';
 void main() {
   runApp(
     GetMaterialApp(
-      title: "SearchAppBarPage",
+      title: 'SearchAppBarPage',
       initialRoute: AppPages.INITIAL,
       getPages: AppPages.routes,
     ),
@@ -19,8 +19,8 @@ class AppPages {
 
   static final routes = [
     GetPage(name: Routes.HOME, page: () => HomePage()),
-    GetPage(name: Routes.PAGE_1, page: () => SearchPage()),
-    GetPage(name: Routes.PAGE_2, page: () => SearchAppBarStream()),
+    GetPage(name: Routes.PAGE_1, page: () => SearchAppBarStream()),
+    GetPage(name: Routes.PAGE_2, page: () => SearchPage()),
   ];
 }
 
@@ -42,7 +42,7 @@ class HomePage extends StatelessWidget {
                   Get.toNamed(Routes.PAGE_1);
                 },
                 child: Text(
-                  'Ir para SearchPage',
+                  'Ir para SearchStreamPage',
                   style: TextStyle(fontSize: 20),
                 )),
             MaterialButton(
@@ -50,7 +50,7 @@ class HomePage extends StatelessWidget {
                   Get.toNamed(Routes.PAGE_2);
                 },
                 child: Text(
-                  'Ir para SearchStreamPage',
+                  'Ir para SearchPage',
                   style: TextStyle(fontSize: 20),
                 )),
           ],
@@ -86,6 +86,8 @@ class SearchPage extends StatelessWidget {
       //listFull: dataList, // Lista String
       listFull: dataListPerson2,
       stringFilter: (Person person) => person.name,
+
+      /// Caso queira sort escolha como fazer
       compareSort: (Person a, Person b) => a.name.compareTo(b.name),
       filtersType: FiltersTypes.contains,
       listBuilder: (list, isModSearch) {
@@ -125,7 +127,15 @@ class SearchPage extends StatelessWidget {
   }
 }
 
-class SearchAppBarStream extends StatelessWidget {
+// ignore: must_be_immutable
+class SearchAppBarStream extends StatefulWidget {
+  const SearchAppBarStream({Key key}) : super(key: key);
+
+  @override
+  _SearchAppBarStreamState createState() => _SearchAppBarStreamState();
+}
+
+class _SearchAppBarStreamState extends State<SearchAppBarStream> {
   @override
   Widget build(BuildContext context) {
     return SearchAppBarPageStream<Person>(
@@ -144,44 +154,69 @@ class SearchAppBarStream extends StatelessWidget {
       listBuilder: (list, isModSearch) {
         // Rertorne seu widget com a lista para o body da page
         // Pode alterar a tela relacionando o tipo de procura
-        return ListView.builder(
-          itemCount: list.length,
-          itemBuilder: (_, index) {
-            return Card(
-                margin: const EdgeInsets.symmetric(horizontal: 16, vertical: 4),
-                shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(4)),
-                // color: Theme.of(context).primaryColorDark,
-                child: Padding(
-                  padding: const EdgeInsets.all(14.0),
-                  child: Row(
-                    children: [
-                      Expanded(
-                        child: Text(
-                          'Name: ${list[index].name}',
-                          style: TextStyle(fontSize: 16),
+        return Column(
+          children: [
+            Expanded(
+              child: ListView.builder(
+                itemCount: list.length,
+                itemBuilder: (_, index) {
+                  return Card(
+                      margin: const EdgeInsets.symmetric(
+                          horizontal: 16, vertical: 4),
+                      shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(4)),
+                      // color: Theme.of(context).primaryColorDark,
+                      child: Padding(
+                        padding: const EdgeInsets.all(14.0),
+                        child: Row(
+                          children: [
+                            Expanded(
+                              child: Text(
+                                'Name: ${list[index].name}',
+                                style: TextStyle(fontSize: 16),
+                              ),
+                            ),
+                            Expanded(
+                              child: Text(
+                                'Age: ${list[index].age.toStringAsFixed(2)}',
+                                style: TextStyle(fontSize: 12),
+                              ),
+                            )
+                          ],
                         ),
-                      ),
-                      Expanded(
-                        child: Text(
-                          'Age: ${list[index].age.toStringAsFixed(2)}',
-                          style: TextStyle(fontSize: 12),
-                        ),
-                      )
-                    ],
-                  ),
-                ));
-          },
+                      ));
+                },
+              ),
+            ),
+            MaterialButton(
+                onPressed: () {
+                  Get.toNamed(Routes.PAGE_2);
+                },
+                child: Text(
+                  'Ir para SearchPage',
+                  style: TextStyle(fontSize: 20),
+                )),
+            MaterialButton(
+                onPressed: () {
+                  setState(() {});
+                },
+                child: Text(
+                  'SetState',
+                  style: TextStyle(fontSize: 20),
+                )),
+          ],
         );
       },
     );
   }
 
   Stream<List<Person>> _streamListPerson = (() async* {
-    await Future<void>.delayed(Duration(seconds: 5));
+    await Future<void>.delayed(Duration(seconds: 4));
     yield dataListPerson;
-    await Future<void>.delayed(Duration(seconds: 10));
+    await Future<void>.delayed(Duration(seconds: 8));
     yield dataListPerson2;
+    await Future<void>.delayed(Duration(seconds: 6));
+    yield dataListPerson3;
   })();
 }
 
@@ -211,6 +246,28 @@ final dataListPerson2 = <Person>[
   Person(name: 'Raquel Montenegro', age: 19),
   Person(name: 'Rafael Peireira', age: 15),
   Person(name: 'Nome Comum', age: 33),
+];
+
+final dataListPerson3 = <Person>[
+  Person(name: 'Rafaela Pinho', age: 30),
+  Person(name: 'Paulo Emilio Silva', age: 45),
+  Person(name: 'Pedro Gomes', age: 18),
+  Person(name: 'Orlando Guerra', age: 23),
+  Person(name: 'Zacarias Triste', age: 15),
+  Person(name: 'Antonio Rabelo', age: 33),
+  Person(name: 'Leticia Maciel', age: 47),
+  Person(name: 'Patricia Oliveira', age: 19),
+  Person(name: 'Pedro Lima', age: 15),
+  Person(name: 'Junior Rabelo', age: 33),
+  Person(name: 'Lucia Maciel', age: 47),
+  Person(name: 'Ana Oliveira', age: 19),
+  Person(name: 'Thiago Silva', age: 33),
+  Person(name: 'Charles Ristow', age: 47),
+  Person(name: 'Raquel Montenegro', age: 19),
+  Person(name: 'Rafael Peireira', age: 15),
+  Person(name: 'Thiago Ferreira', age: 33),
+  Person(name: 'Joaquim Gomes', age: 18),
+  Person(name: 'Esther Guerra', age: 23),
 ];
 
 class Person {