+++ lib/src/search_app_bar_page/controller/seacher_base_controll.dart
@@ -7,7 +7,7 @@ abstract class SeacherBase{
 
   set isModSearch(bool value) => _isModSearch.value = value;
 
-  final rxSearch = ''.obs;
+  final RxString rxSearch = ''.obs;
 
   final RxBool _bancoInit = false.obs;