+++ lib/src/search_app_bar_page/filters/filter.dart
@@ -1,38 +1,27 @@
 import 'package:diacritic/diacritic.dart';
 import 'package:flutter/material.dart';
 
-//enum ComparesTypes { string, number }
+typedef Filter<T> = bool Function(T test, String query);
 
-typedef bool Filter<T>(T test, String query);
+typedef StringFilter<T> = String Function(T test);
 
-typedef String StringFilter<T>(T test);
+typedef Compare<T> = int Function(T a, T b);
 
-typedef int Compare<T>(T a, T b);
-
-typedef bool Funcion(bool isModSearch);
-
-typedef Widget FunctionList<T>(List<T> list, bool isModSearch);
-
+typedef FunctionList<T> = Widget Function(List<T> list, bool isModSearch);
 
 class Filters {
-  /// returns if [test] starts with the given [query],
-  /// disregarding lower/upper case and diacritics.
   static Filter<String> startsWith = (test, query) {
     final realTest = _prepareString(test);
     final realQuery = _prepareString(query);
     return realTest.startsWith(realQuery);
   };
 
-  /// returns if [test] is exactly the same as [query],
-  /// disregarding lower/upper case and diacritics.
   static Filter<String> equals = (test, query) {
     final realTest = _prepareString(test);
     final realQuery = _prepareString(query);
     return realTest == realQuery;
   };
 
-  /// returns if [test] contains the given [query],
-  /// disregarding lower/upper case and diacritics.
   static Filter<String> contains = (test, query) {
     final realTest = _prepareString(test);
     final realQuery = _prepareString(query);