+++ lib/src/search_app_bar_page/search_app_bar_page_stream.dart
@@ -1,12 +1,15 @@
 import 'dart:async';
+
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
+
+import 'controller/searcher_page_stream_controller.dart';
 import 'filters/filter.dart';
 import 'filters/filters_type.dart';
 import 'search_app_bar/search_app_bar.dart';
-import 'controller/searcher_page_stream_controller.dart';
 import 'stream_helper/stream_search_builder.dart';
 
+// ignore: must_be_immutable
 class SearchAppBarPageStream<T> extends StatefulWidget {
   /// Paramentros do SearchAppBar
 
@@ -21,10 +24,21 @@ class SearchAppBarPageStream<T> extends StatefulWidget {
   final TextCapitalization searchAppBarcapitalization;
   final List<Widget> searchAppBaractions;
   final double searchAppBarelevation;
+  final bool showIconConnectyOffAppBar;
+  final Widget iconConnectyOffAppBar;
+
+  ///  [iconConnectyOffAppBar] Aparece quando o status da conexao é off.
+  ///  já existe um icone default. Caso nao queira apresentar escolha
+  ///  [showIconConnectyOffAppBar] = false;
+
 
   /// Parametros para o Scaffold
 
+  ///  [widgetConnecty] Apenas mostra algo quando esta sem conexao e ainda nao tem o primeiro
+  /// valor da stream. Se a conexao voltar passa a mostrar o [widgetWaiting] até apresentar
+  /// o primeiro dado
   final Widget widgetWaiting;
+  final Widget widgetConnecty;
   final Widget widgetError;
   final Widget searchePageFloaActionButton;
   final FloatingActionButtonLocation searchePageFloatingActionButtonLocation;
@@ -55,7 +69,9 @@ class SearchAppBarPageStream<T> extends StatefulWidget {
   final Compare<T> compareSort;
 
   /// Para montar em asBroadcastStream
-  final Stream<List<T>> _stream;
+  // nao de pode colocar final. Após um setState precisa refaze-la
+  // vide o metodo didUpdateWidget
+  Stream<List<T>> _stream;
 
   SearchAppBarPageStream({
     Key key,
@@ -66,6 +82,7 @@ class SearchAppBarPageStream<T> extends StatefulWidget {
     this.compareSort,
     this.filtersType,
     this.stringFilter,
+    this.widgetConnecty,
 
     /// Paramentros do SearchAppBar
     this.searchAppBartitle,
@@ -79,6 +96,8 @@ class SearchAppBarPageStream<T> extends StatefulWidget {
     this.searchAppBarcapitalization = TextCapitalization.none,
     this.searchAppBaractions = const <Widget>[],
     this.searchAppBarelevation = 4.0,
+    this.showIconConnectyOffAppBar = true,
+    this.iconConnectyOffAppBar,
 
     /// Parametros para o Scaffold
 
@@ -104,7 +123,8 @@ class SearchAppBarPageStream<T> extends StatefulWidget {
     this.drawerEdgeDragWidth,
     this.drawerEnableOpenDragGesture = true,
     this.endDrawerEnableOpenDragGesture = true,
-  })  : _stream = listStream.asBroadcastStream(),
+  })
+      : _stream = listStream.asBroadcastStream(),
         super(key: key);
 
   @override
@@ -113,24 +133,22 @@ class SearchAppBarPageStream<T> extends StatefulWidget {
 }
 
 class _SearchAppBarPageStreamState<T> extends State<SearchAppBarPageStream<T>> {
-  SearcherPageStreamController _controller;
+  SearcherPageStreamController<T> _controller;
+
+  StreamSubscription _streamSubscription;
 
   @override
   void initState() {
     super.initState();
     _controller = SearcherPageStreamController<T>(
-        listStream: widget._stream,
+      //listStream: widget._stream,
         stringFilter: widget.stringFilter,
         compareSort: widget.compareSort,
         filtersType: widget.filtersType)
       ..onInit()
-      ..onReady();
-  }
+      ..subscribeWorker();
 
-  @override
-  void dispose() {
-    _controller.onClose();
-    super.dispose();
+    _subscribeListStream();
   }
 
   @override
@@ -149,9 +167,12 @@ class _SearchAppBarPageStreamState<T> extends State<SearchAppBarPageStream<T>> {
           flattenOnSearch: widget.searchAppBarflattenOnSearch,
           capitalization: widget.searchAppBarcapitalization,
           actions: widget.searchAppBaractions,
+          showIconConnectyOffAppBar: widget.showIconConnectyOffAppBar,
+          iconConnectyOffAppBar: widget.iconConnectyOffAppBar,
         ),
         body: StreamSearchBuilder<T>(
             initialData: widget.initialData,
+            widgetConnecty: widget.widgetConnecty,
             stream: widget._stream,
             searcher: _controller,
             listBuilder: widget.listBuilder,
@@ -160,9 +181,8 @@ class _SearchAppBarPageStreamState<T> extends State<SearchAppBarPageStream<T>> {
                 if (widget.widgetError == null)
                   return Column(
                       mainAxisAlignment: MainAxisAlignment.center,
-                      crossAxisAlignment: CrossAxisAlignment.center,
                       children: <Widget>[
-                        Icon(
+                        const Icon(
                           Icons.error_outline,
                           color: Colors.red,
                           size: 60,
@@ -179,10 +199,16 @@ class _SearchAppBarPageStreamState<T> extends State<SearchAppBarPageStream<T>> {
               } else {
                 if (widget.widgetWaiting == null)
                   return Center(
-                    child: SizedBox(
-                      child: const CircularProgressIndicator(),
-                      width: 60,
-                      height: 60,
+                    child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: const [
+                        SizedBox(
+                          width: 60,
+                          height: 60,
+                          child: CircularProgressIndicator(),
+                        ),
+                      ],
                     ),
                   );
                 else
@@ -191,9 +217,9 @@ class _SearchAppBarPageStreamState<T> extends State<SearchAppBarPageStream<T>> {
             }),
         floatingActionButton: widget.searchePageFloaActionButton,
         floatingActionButtonLocation:
-            widget.searchePageFloatingActionButtonLocation,
+        widget.searchePageFloatingActionButtonLocation,
         floatingActionButtonAnimator:
-            widget.searchePageFloatingActionButtonAnimator,
+        widget.searchePageFloatingActionButtonAnimator,
         persistentFooterButtons: widget.searchePagePersistentFooterButtons,
         drawer: widget.searchePageDrawer,
         endDrawer: widget.searchePageEndDrawer,
@@ -210,6 +236,52 @@ class _SearchAppBarPageStreamState<T> extends State<SearchAppBarPageStream<T>> {
         drawerEdgeDragWidth: widget.drawerEdgeDragWidth,
         drawerEnableOpenDragGesture: widget.drawerEnableOpenDragGesture,
         endDrawerEnableOpenDragGesture: widget.endDrawerEnableOpenDragGesture);
+  }
+
+  void _subscribeListStream() {
+    _streamSubscription = widget._stream.listen((listData) {
+      if (_controller.bancoInit) {
+        // Fica negativo dentro do StreamBuilder
+        // Após apresentar o primeiro Obx(())
+        _controller.listFull = listData;
+        if (_controller.rxSearch.value.isNotEmpty) {
+          _controller.refreshSeachList(_controller.rxSearch.value);
+        } else {
+          _controller.sortCompareList(listData);
+          _controller.onSearchFilter(listData);
+        }
+      } else {
+        _controller.initialChangeList = listData;
+      }
+    });
+  }
+
+  void _unsubscribeListStream() {
+    if (_streamSubscription != null) {
+      _streamSubscription.cancel();
+      _streamSubscription = null;
+    }
+  }
+
+  @override
+  void didUpdateWidget(SearchAppBarPageStream<T> oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (oldWidget._stream != widget._stream) {
+      if (_streamSubscription != null) {
+        _unsubscribeListStream();
+        //Necessário para resolver após o setState
+        widget._stream = oldWidget._stream.asBroadcastStream();
+      }
+
+      _subscribeListStream();
+    }
+  }
+
+  @override
+  void dispose() {
+    _controller.onClose();
+    _unsubscribeListStream();
+    super.dispose();
   }
 }