+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.1.0]
+- Agora voce pode usar sua SearchAppBarPageStream dentro de um StatefulWidget e dar setState.
+- Adicionei um mecanismo de verificar a conexao e mostra um widget [widgetConnecty] na primeira requisicao de dados. Aparece quando nao existe dado e a conexao esta off. Tambem no App Bar já existe um icone de status off por default. Voce pode mudar a cor, o tipo ou nao mostrar.
+
 ## [0.0.2]
 - Correçoes no read e na example.